src/components: move BannerAppButtons
